fix user update

+++ controllers/userController.js
@@ -78,7 +78,7 @@ const login = async (req, res) => {
   await User.findByIdAndUpdate(user.id, { accessToken, refreshToken });
 
   res.json({
-    token:accessToken,
+    token: accessToken,
     refreshToken,
     user: {
       id: user.id,
@@ -140,17 +140,21 @@ const getCurrent = async (req, res) => {
 const update = async (req, res) => {
   const { _id } = req.user;
   const { name, email, phone, skype, birthday } = req.body;
-  const { fieldname, path: oldPath } = req.file;
-  const fileData = await cloudinary.uploader.upload(oldPath, {
-    folder: "avatar",
-  });
-  await fs.unlink(oldPath);
 
   const fieldsToUpdate = {};
 
-  if (fieldname === "avatarURL") {
-    fieldsToUpdate.avatarURL = fileData.url;
+  if (req.file) {
+    const { fieldname, path: oldPath } = req.file;
+    const fileData = await cloudinary.uploader.upload(oldPath, {
+      folder: "avatar",
+    });
+    await fs.unlink(oldPath);
+
+    if (fieldname === "avatarURL") {
+      fieldsToUpdate.avatarURL = fileData.url;
+    }
   }
+
   if (name) {
     fieldsToUpdate.name = name;
   }
@@ -166,8 +170,6 @@ const update = async (req, res) => {
   if (birthday) {
     fieldsToUpdate.birthday = birthday;
   }
-
-  console.log(fieldsToUpdate);
 
   const result = await User.findByIdAndUpdate(_id, fieldsToUpdate, {
     new: true,
